Flask_API_AWS/pyproject.toml
[tool.poetry]
name = "flask_api_aws"
version = "0.1.0"
description = "AzePUG live stream building REST API with AWS"
authors = ["Shahriyar Rzayev <rzayev.sehriyar@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
img2pdf = ">=0.4.4"
Pillow = ">=9.1.0"

[tool.poetry.dev-dependencies]
pytest = ">=7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
